# Conflicts:
#	ISY-SIM/isy/sim/FXMLSetUpController.java

+++ ISY-SIM/isy/sim/Layers/LandLayer.java
@@ -31,7 +31,7 @@ import javafx.scene.canvas.GraphicsContext;
  */
 public class LandLayer extends Layer {
 
-    public LandLayer(GraphicsContext gContext, int width, int height) {
+    public LandLayer(GraphicsContext gContext, double width, double height) {
         super(gContext, width, height);
     }
