+++ ISY-SIM/isy/sim/Layers/Layer.java
@@ -25,6 +25,7 @@
 package sim.Layers;
 
 import javafx.scene.canvas.GraphicsContext;
+import javafx.scene.paint.Color;
 import sim.Utilities.SimMatrix;
 
 /**
@@ -35,9 +36,9 @@ public abstract class Layer {
     GraphicsContext gc;
     public SimMatrix m;
 
-    public Layer(GraphicsContext gContext, int width, int height) {
+    public Layer(GraphicsContext gContext, double width, double height) {
         gc = gContext;
-        m = new SimMatrix(width, height);
+        m = new SimMatrix((int)width, (int)height);
     }
 
     public void drawLayer() {
